pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "northstar"
version = "0.1.0"
description = "The Priority Debt Engine — measures the cost of working on low-leverage tasks"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "NorthStar Team"}]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "anthropic>=0.25.0",
    "aiosqlite>=0.19.0",
    "pathspec>=0.11.0",
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]
ast = [
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
]
anthropic = [
    "strands-agents[anthropic]>=0.1.0",
]

[project.scripts]
northstar = "northstar.cli:app"

[tool.setuptools.packages.find]
include = ["northstar*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
